+++ src/main/java/com/hadoop/relativefrequency/pairs/RelativeFrequencyTool.java
@@ -4,7 +4,7 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.input.TextInputFormat;
@@ -42,7 +42,7 @@ public class RelativeFrequencyTool extends Configured implements Tool {
 
 		// Specify key / value
 		job.setOutputKeyClass(WordPairKeyObject.class);
-		job.setOutputValueClass(IntWritable.class);
+		job.setOutputValueClass(DoubleWritable.class);
 
 		// Input
 		FileInputFormat.addInputPath(job, inputPath);
